Flyttet sjablonPeriode og Avvik fra bidragsevne til felles
