+++ src/main/java/no/nav/bidrag/beregn/felles/bidragsevne/periode/BidragsevnePeriode.java
@@ -2,7 +2,7 @@ package no.nav.bidrag.beregn.felles.bidragsevne.periode;
 
 import java.util.List;
 import no.nav.bidrag.beregn.felles.bidragsevne.beregning.Bidragsevneberegning;
-import no.nav.bidrag.beregn.felles.bidragsevne.bo.Avvik;
+import no.nav.bidrag.beregn.felles.bo.Avvik;
 import no.nav.bidrag.beregn.felles.bidragsevne.bo.BeregnBidragsevneGrunnlagAlt;
 import no.nav.bidrag.beregn.felles.bidragsevne.bo.BeregnBidragsevneResultat;
